requirements.txt
boto==2.46.1
Collectfast==0.5.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.6
django-appconf==1.0.2
django-compressor==2.1.1
django-cors-headers==2.2.0
django-debug-toolbar==1.7
django-hide-herokuapp==0.4
django-ipware==1.1.6
django-storages==1.5.2
django-storages-redux==1.3.2
django-toolbelt==0.0.1
gunicorn==19.7.0
mock==2.0.0
newrelic==2.80.1.61
pbr==4.0.2
psycopg2==2.7.4
python-dateutil==2.4.2
pytz==2016.10
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.3
static3==0.7.0
tzlocal==1.2
Unidecode==0.4.18
